requirements.txt
attrs==25.3.0
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==6.0.0
packaging==25.0
parsel==1.10.0
Protego==0.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
python-dotenv==1.1.1
queuelib==1.8.0
requests==2.32.4
requests-file==2.1.0
Scrapy==2.13.3
service-identity==24.2.0
tldextract==5.3.0
tomli==2.2.1
Twisted==25.5.0
typing_extensions==4.14.1
urllib3==2.5.0
w3lib==2.3.1
zope.interface==7.2
